CHG: move enums.py, types.py
