+++ promptmodel/database/crud.py
@@ -12,7 +12,7 @@ from promptmodel.database.models import (
     DeployedPrompt,
 )
 from playhouse.shortcuts import model_to_dict
-from promptmodel.utils.enums import ModelVersionStatus, ParsingType
+from promptmodel.types.enums import ModelVersionStatus, ParsingType
 from promptmodel.utils.random_utils import select_version
 from promptmodel.utils import logger
 from promptmodel.database.config import db
