+++ promptmodel/database/models.py
@@ -18,7 +18,7 @@ from peewee import (
 
 from promptmodel.database.config import BaseModel
 from promptmodel.database.models_chat import *
-from promptmodel.utils.enums import ModelVersionStatus, ParsingType
+from promptmodel.types.enums import ModelVersionStatus, ParsingType
 
 
 class JSONField(TextField):
