+++ tests/cli/dev_test.py
@@ -6,4 +6,4 @@ from litellm import ModelResponse
 
 from promptmodel.llms.llm import LLM
 from promptmodel.llms.llm_proxy import LLMProxy
-from promptmodel.utils.types import LLMResponse, LLMStreamResponse
+from promptmodel.types.response import LLMResponse, LLMStreamResponse
