+++ tests/llm/parse_test.py
@@ -8,8 +8,8 @@ from litellm import ModelResponse
 
 from promptmodel.llms.llm import LLM
 from promptmodel.llms.llm_proxy import LLMProxy
-from promptmodel.utils.types import LLMResponse, LLMStreamResponse
-from promptmodel.utils.enums import ParsingPattern, ParsingType, get_pattern_by_type
+from promptmodel.types.response import LLMResponse, LLMStreamResponse
+from promptmodel.types.enums import ParsingPattern, ParsingType, get_pattern_by_type
 
 
 def generator_format(response: str):
