+++ tests/llm/stream_test.py
@@ -7,7 +7,7 @@ from litellm import ModelResponse
 
 from promptmodel.llms.llm import LLM
 from promptmodel.llms.llm_proxy import LLMProxy
-from promptmodel.utils.types import LLMResponse, LLMStreamResponse
+from promptmodel.types.response import LLMResponse, LLMStreamResponse
 from promptmodel.utils.prompt_util import run_async_in_sync
